adding new pages, not connected to the routing yet
