+++ client/src/js/AuthPage.jsx
@@ -1,19 +1,21 @@
-import React from 'react';
-import {useLocation} from 'react-router-dom';
+import React, { useState } from 'react';
+import {Link} from 'react-router-dom';
 import PageContainer from './PageContainer';
+import { Button, Modal } from 'react-bootstrap';
+import 'bootstrap/dist/css/bootstrap.min.css';
 
 
 const AuthComponent = () => {
   return (
-    <div>
-        <h2>History</h2>
+    <div className="module authpage green-2">
+        <h2>Log In</h2>
     </div>
-  )
+  );
 }
 
 const AuthPage = () => {
   return (
-    <PageContainer hasHeader className="FormPage">
+    <PageContainer>
       <AuthComponent />
     </PageContainer>
   )
